tst_QWidget: fix a memleak in destroyedSignal()
The test checks that the destroyed() signal is emitted, even in
exceptional cases. It has three degrees of freedom:
- blockSignals() or not
- QObject* or QWidget* as the connect() sender
- slot in a child of the deleted object or in an unrelated object
  (one of the latter caused a leak of the DestroyTester)

The code was very repetitive and one scenario may have been checked
twice (QObject*, no child, blocked signals), which probably means we
forgot another (didn't check).

Make the test data-driven, and DRY further using templates for the
QObject/QWidget cases.

The leak fix is to use unique_ptr to hold the QWidget, and a scope
guard to delete the DestroyTester. I didn't want to use QPointer (or
qdialog_p's QAutoPointer), because we're checking a mechanism that
strikes very close to QPointer, and we should not depend on knowledge
of the exact QPointer implementation to convince ourselves that we're
not testing the same thing twice here. Besides, by making the delete t
conditional, we can use ASAN to detect if we got the need to delete
wrong.

Amends e72e5aa83a0a287ea7e19158b39917eb3daef1d9.

Pick-to: 6.9 6.8 6.5 5.15
Change-Id: I3bac9f81c33a24e1096041dfc3323e079fd46f8a
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>