QGIM: add MultiColumn wrapper for disambiguation
If a type has both a metaobject and implements the tuple protocol, then
we need to disambiguate as row_traits would be ambiguous.

Add  a wrapper that holds the type and provides a forwarding tuple
protocol implementation. With either or the SingleColumn wrapper, client
code can decide whether the type should be a a multi-column row in a
table (accessed via tuple protocol), or  a multi-role item in a list
(accessed via the meta object).

Add documentation.

Change-Id: I0145f7a35a125a138254397bc7eafa5506e5e3ad
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>