QGIM: test if a tuple row is valid before dereferencing
Row types that are convertible to bool, such as pointers, but also
unique_ptr and similar wrappers, should not be dereferenced when
accessing values for reading or writing. This is already done implicitly
when operating on single-value types, but for tuples we have to do so
explicitly, before accessing the element.

For example, inserting a row where the row type is a pointer will leave
that row be nullptr in the range, and both data and setData should fail
until a row has been constructed. For rows that are pointers to tuples,
we must never call get<>, as get cannot test if the item received is a
nullptr and still return a valid value (reference).

Since e.g. unique_ptr is explicitly convertible to bool we have to use
std::is_constructible rather than is_convertible, as the latter only
works for types that are implicitly convertible.

Change-Id: I9dbdf6e5cac0146ed5d1b88c523e182590b0c8ab
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>