QFactoryLoader: drop libraries mutex sooner
Both QFactoryLoader::metaDataKeys() and QFactoryLoader::metaData()
locked the mutex under QT_CONFIG(library) near the start of the
function and never dropped it until the end of the function.

This is not necessary and, by Amdahl's Law, reduces concurrency.

Drop the mutex before the #endif again.

This isn't as good as splitting the metaDataKeys() loop into one that
extracts the QPluginParsedMetaDatas, dropping the mutex, and then
another doing the JSON work, but it's better than the status quo and
the best the maintainer will accept.

Amends 878e3342e1c9bd8a4da6a2e9e1508dacbe0c7ab6.

Pick-to: 6.10 6.9
Change-Id: Ie69d4a2d20ac02c4331ffcc3f1281caf7cbeb6fe
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>