Add constant property QStyleHints::accessibility
This property bundles together accessibility settings that can be used
by styles to change how they draw different widgets and/or controls.

The only property that it contains currently is contrastPreference.
This property can be used to determine if the user has enabled a high
contrast setting on the platform that the Qt application is running on.

We might add more *accessibility* settings, like color blind hints, to
QAccessibilityHints in the future.

[ChangeLog][Gui][QStyleHints] Added class QAccessibilityHints, which can be
accessed from QStyleHints. This singleton contains a single property
contrastPreference which can be used to determine if the platform has
enabled a high contrast setting or not. More accessibility settings
might be added to QAccessibilityHints in the future.

Task-number: QTBUG-133595
Change-Id: I93c6be81d29f65058eeda00328f4d9e966921de6
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>