tst_QCalendarWidget: fix memleak in showPrevNext()
QObject parents own their children, but not the other way around. So
the parent-less QWidget *parent created on the heap was leaked.

To fix, create it on the stack instead, but for that to work, it must
be declared _before_ its eventual child, otherwise the parent would
delete the child as a QObject child, followed by the compiler deleting
the child as an automatic variable = double-delete.

Amends ab536c3c71deaaa3b67ca87c47628355d9f348e8.

Pick-to: 6.9 6.8 6.5 5.15
Change-Id: Ibccf229fc056665e3b854e1a94e1f4b5bf7a89e1
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>