Fix qRound(inf) assert
In some extreme cases we can return a finite float, and then turn it
infinite when multiplying it to calculate the index positions.

Avoid the whole thing by clamping to a brightness of +-32768 times the
white point standard, which is well within current HDR standards.

Credit to OSS-Fuzz which detected the assert.

Pick-to: 6.9 6.8
Change-Id: If007732a8d59ea27514f17674d318a099a057281
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Robert Löhning <robert.loehning@qt.io>