CMake: Allow echoing RunCMake test command output via an opt-in
This will be used in a new manual test in qt5.git, to show progress of
long-running commands.

Store the diff in a new patches subdir, to make it easier to re-apply
when we update the file from upstream.

As a drive-by, store patches from the previous modifications that were
done before this change.

Pick-to: 6.8 6.9
Change-Id: Ib73908e20fac7fe0bd620400d483625238fa37d1
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Cristian Le <cristian.le@qt.io>