tst_QSplitter: don't leak the QSplitter from initTestCase()
Several test functions use a centrally-allocated QSplitter for their
tests. This object was allocated in initTestCase(), but
cleanTestCase(), while present, was empty, so nothing deleted the
object again.

Fix by deleting the object in cleanupTestCase(). Initialize the member
variable to nullptr so that the delete is well-formed even if
initTestCase() for some reason didn't run (or didn't run to
completion).

Amends the start of the public history, where this function read

    #if defined(QT3_SUPPORT)
        delete qApp->mainWidget();
    #endif

which was already leaking if QT3_SUPPORT wasn't defined.

Pick-to: 6.9 6.8 6.5 5.15
Change-Id: I43e5d8921a9c097f252e11709f25bc622fc8fe15
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>