QGIM: check that row is valid also in flags()
While we never access the data, accessing a MultiColumn item holding a
nullptr would still dereference that nullptr, and then pass that
reference to our lambda.

And the results are misleading: if there is no gadget or object stored
for that row yet, then the item is not editable - nothing can be stored.

Change-Id: Ic24242296f8cdedae6e36a12da470831eb93ee2a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>