GnomeTheme: read the initial color-scheme value from dbus
The initial color-scheme value was never initialized, and would thus,
always be 'Unknown', until the user changes the color scheme in the
system settings.

The xdg settings portal method ReadAll is currently omitting the
contrast setting from the reply. Because of this, two separate dbus
calls are necessary to get both the initial color-scheme and contrast
setting.

Change-Id: Ie32371efd830e19727922efd2c9c32595395d11a
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>