QVariant: fix assert in FP->int conversions
The newly introduced asserts in qRound make QVariant FP->int conversions
crash. The crash is legitimate, because the conversion has UB. However
we've historically supported these conversions with a somehow specific
(albeit undocumented) behavior; I'm trying to keep it as accurately as
possible, based on testing on x86-64. (I have no idea if the behavior
was actually different elsewhere. Formally there's no "behavior" to
talk about here.)

Specifically: use "saturation rounding" towards qint64, and then convert
the result to the target type. This can be done by using the private
qSaturateRound function, which however needs to be extended to 64 bits.

This extension revealed an implementation problem: using qBound to clamp
the result of the rounding is inaccurate for certain combination of
input and output types (e.g. double to qint64), because the maximal
value can't be correctly represented in floating point.  Use a better
algorithm instead.

Finally, I'm making NaN *fail* the conversion (and return 0). I don't
think there's any other sensible alternative here. I'm not sure of the
pre-existing behavior of the code across different architectures and
using different encodings for NaN, and it sounds extremely brittle to do
something else instead.

Fixes: QTBUG-135285
Change-Id: Ia5f54094807210d0dbbdbb9a0a4f3a0f18ea3d80
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>